NPC work  Saves  Swamp Level
Reworked the NPC class to remove unneeded subclasses. Made it possible
to move an NPC away from a blocked location and for the save function to
save and load the new location of moved NPCs. Referrals can be saved.
Started work on the swamp level and made an NPC to interact with.